MetaTransformer-Scripts/quantum-poetry-system/requirements.txt
# Core
numpy>=1.21.0
openai>=0.27.0
python-dotenv>=1.0.0

# For resonance patterns
scipy>=1.7.0

# Development
pytest>=7.0.0
black>=22.0.0